Val_Projects/cpu_mem_stress_remote/cpu_mem_stress_remote/requirements.txt
paramiko
pyyaml
pytest
